backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2 
psycopg2-binary==2.9.9
psycopg2==2.9.9
django-allauth==0.63.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# 2FA
django-two-factor-auth==1.15.5
phonenumbers==8.13.44
django-otp==1.5.2

# Dependencies for django-alliauth
requests
cryptography
